Create order page for cook
1. Create order page include view and html.
2. Move testing helper method into test-base.py
