+++ CookManagement/templates/menu.html
@@ -1,41 +1,41 @@
 {% extends 'base.html' %}
 
 {% block page %}
-<div class="panel">
-  <div class="panel-heading bg-blue">
-    <h4 class="panel-title text-center text-white">Menu</h4>
-  </div>
-  <div class="panel-body">
-    <div class="text-right">
-      <a href="{% url 'cook-add-meal' %}" class="btn btn-pink">Add Meal</a>
-      <br/>
-      <br/>
+  <div class="panel">
+    <div class="panel-heading bg-blue">
+      <h4 class="panel-title text-center text-white">Menu</h4>
     </div>
+    <div class="panel-body">
+      <div class="text-right">
+        <a href="{% url 'cook-add-meal' %}" class="btn btn-pink">Add Meal</a>
+        <br/>
+        <br/>
+      </div>
 
-    <table class="table table-bordered table-hover table-striped">
-      <thead>
-        <tr class="bg-gray text-white">
-          <th>Id</th>
-          <th>Name</th>
-          <th>Description</th>
-          <th>Price</th>
-          <th class="text-center">Image</th>
-        </tr>
-      </thead>
-      <tbody>
-        {% for meal in meals %}
-          <tr>
-            <td scope="row">{{ meal.id }}</td>
-            <td><a href="{% url 'cook-edit-meal' meal.id %}">{{ meal.Name }}</a> </td>
-            <td>{{ meal.Description }}</td>
-            <td>{{ meal.Price }}</td>
-            <td class="text-center"><img src="{{ MEDIA_URL }}{{meal.Image}}"
-                class="img-rounded" width="50" height="50"></td>
+      <table class="table table-bordered table-hover table-striped">
+        <thead>
+          <tr class="bg-gray text-white">
+            <th>Id</th>
+            <th>Name</th>
+            <th>Description</th>
+            <th>Price</th>
+            <th class="text-center">Image</th>
           </tr>
-        {% endfor%}
-      </tbody>
-    </table>
+        </thead>
+        <tbody>
+          {% for meal in meals %}
+            <tr>
+              <td scope="row">{{ meal.id }}</td>
+              <td><a href="{% url 'cook-edit-meal' meal.id %}">{{ meal.Name }}</a> </td>
+              <td>{{ meal.Description }}</td>
+              <td>{{ meal.Price }}</td>
+              <td class="text-center"><img src="{{ MEDIA_URL }}{{meal.Image}}"
+                  class="img-rounded" width="50" height="50"></td>
+            </tr>
+          {% endfor %}
+        </tbody>
+      </table>
 
+    </div>
   </div>
-</div>
 {% endblock %}
