+++ CookManagement/templates/order.html
@@ -1,4 +1,55 @@
 {% extends 'base.html' %}
 
 {% block page %}
+  <div class="panel">
+    <div class="panel-heading bg-blue">
+      <h4 class="panel-title text-center text-white">Orders</h4>
+    </div>
+    <div class="panel-body">
+      <table class="table table-bordered table-hover table-striped">
+        <thead>
+          <tr class="bg-gray text-white">
+            <th>Id</th>
+            <th>Order Details</th>
+            <th>Customer</th>
+            <th>Total Price</th>
+            <th>Status</th>
+            <th>Action</th>
+          </tr>
+        </thead>
+        <tbody>
+          {% for order in orders %}
+            <tr>
+              <td scope="row" >{{order.id}}</td>
+              <!--TODO: Make order details expandable -->
+              <td>
+                {% for detail in order.order_detail.all %}
+                  {{ detail.Meal.Name }}: {{ detail.Meal.Price }} x {{ detail.Quantity }} = ${{ detail.SubtotalPrice }}
+                {% endfor %}
+              </td>
+              <td>{{ order.Customer }}</td>
+              <td>{{ order.TotalPrice }}</td>
+              <td>{{ order.get_Status_display }}</td>
+              <td>
+                {% if order.Status == 0 %}
+                  <form method="POST">
+                    {% csrf_token %}
+                    <input name="id" value="{{ order.id }}" hidden>
+                    <button class="btn btn-success">Accept</button>
+                  </form>
+                {% elif order.Status == 1%}
+                  <form method="POST">
+                    {% csrf_token %}
+                    <input name="id" value="{{ order.id }}" hidden>
+                    <button class="btn btn-success">Ready</button>
+                  </form>
+                {% endif %}
+              </td>
+            </tr>
+          {% endfor %}
+        </tbody>
+      </table>
+
+    </div>
+  </div>
 {% endblock %}
