+++ Identity/admin.py
@@ -1,7 +1,8 @@
 from django.contrib import admin
 
 # Register your models here.
-from Identity.models import Cook, User
+from Identity.models import Cook, User, Customer
 
 admin.site.register(User)
 admin.site.register(Cook)
+admin.site.register(Customer)
