+++ Identity/tests/test_views.py
@@ -154,6 +154,7 @@ class DefaultRegistrationTests(TestBase):
 
     def test_cook_order_page(self):
         self.signup_new_user()
+        self.apply_cook()
 
         resp = self.client.get(reverse(self.COOK_ORDER_URL))
         self.assertEquals(resp.status_code, 200)
@@ -165,6 +166,7 @@ class DefaultRegistrationTests(TestBase):
 
     def test_cook_report_page(self):
         self.signup_new_user()
+        self.apply_cook()
 
         resp = self.client.get(reverse(self.COOD_REPORT_URL))
         self.assertEquals(resp.status_code, 200)
@@ -173,25 +175,3 @@ class DefaultRegistrationTests(TestBase):
 
         resp = self.client.get(reverse(self.COOD_REPORT_URL))
         self.assertRedirects(resp, '/cook/sign-in/?next=/cook/report/')
-
-    # Sign up for a new user
-    def signup_new_user(self):
-        self.client.post(reverse(self.SIGNUP_URL),
-                         data={self.FIELD_EMAIL: self.DEFAULT_EMIAL,
-                               self.FIELD_PASSWORD: self.DEFAULT_PASSWORD})
-
-    # Log in with valid user
-    def user_login(self):
-        self.client.login(username=self.DEFAULT_EMIAL,
-                          password=self.DEFAULT_PASSWORD)
-
-    # Become a cook
-    def apply_cook(self):
-        with open(self.DEFAULT_LOGO, 'rb') as logo:
-            self.client.post(reverse(self.COOK_APPLY_URL),
-                             data={self.FIELD_NAME: self.DEFAULT_NAME,
-                                   self.FIELD_DESC: self.DEFAULT_DESC,
-                                   self.FIELD_TAG: self.DEFAULT_TAG,
-                                   self.FIELD_ADDRESS: self.DEFAULT_ADDRESS,
-                                   self.FIELD_PHONE: self.DEFAULT_PHONE_NUMBER,
-                                   self.FIELD_LOGO: logo})
